Extended test cases, fixed unregister to not send back a silly answer.
If the server does not allow us to unregister a function when we do a mass
removal, we will now be able to keep functioning. If this case happens, I
suspect the user will want to just destroy the worker_st and start all over.

A gearman_sterror() was added so that you can pretty print the return code.
You can also now find out if a function exists in the worker.