Add basic documentation of undocumented protocol
It seems that the protocol added a few more commands and responses in
recent times, and they should at least be documented. The way that REDUCER
is used is still somewhat confusing and needs better documentation as
to what is expected, but at least an implementor can handle the commands
with a graceful error now.