Fix building on MacOS
There were a number of issues that built up over time as we weren't
testing on OS X, and as clang got more pedantic. They are listed below
as best I understand them:

htonll - OS X defines this. Linux does not. We took a few strategies to
get around this:

 - Remove unused htonll defs/calls
 - Do not try to redefine htonll
 - Add HAVE_HTONLL to ifdef around when necessary.

The pragmas in libgearman/command and error_code allow us to keep
building while gperf fixes their issues that clang exposes. We tried
with gperf 3.1, but this only fixed the register issue, not the cast
precision warnings.

A few things were just minor warnings that clang reports that gcc does
not.